+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.1.0
+- Added `Number` class.
+- Added `style` argument to `kanji2int` and `int2kanji` that specifies style of formatting.
+- Added `kanji_thousand` to `int2kanji` that specifies how thousand digits are kanjized.
+
 # v0.0.2
 - fixed bug Returns wrong answer in kanji2int when given number is greater than 10 ** 16